+++ docs/changes.md
@@ -13,6 +13,7 @@
 - add `shapes.ThickTube()` as per #800
 - add `settings.backend_autoclose` (default is True) to automatically close `Plotter` after show 
 in jupyter notebooks, #802
+- fix k3d backend - remove pinning to 2.7.4 #808
 
 
 -------------------------